+++ packages/components/src/fab/index.ts
@@ -1,2 +1,3 @@
 export { FAB } from './FAB'
+export { FAB_ICON_SIZES, FAB_SIZES } from './styles'
 export type { FABProps, FABSize, FABVariant } from './types'